backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Natural Language Processing
nltk>=3.8.1
textblob>=0.17.1

# OpenAI Integration
openai>=1.0.0
tiktoken>=0.5.0

# Environment and Configuration
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0 